requirements.txt
python-telegram-bot==21.0
openai==1.14.0
httpx==0.27.0
python-dotenv==1.0.1
